haciendo otro broken_gnl y cambiando cositas en n_queens

+++ exam_03/broken_gnl/gnl.h
@@ -26,10 +26,10 @@
 
 char	*get_next_line(int fd);
 char	*ft_strchr(char *s, int c);
-size_t	ft_strlen(char *s);
-void	*ft_memcpy(void *dest, const void *src, size_t n);
-int	str_append_mem(char **s1, char *s2, size_t size2);
-int	str_append_str(char **s1, char *s2);
-void	*ft_memmove(void *dest, const void *src, size_t n);
+int		ft_strlen(char *s);
+void	*ft_memcpy(void *dest, const void *src, int n);
+int		str_append_mem(char **s1, char *s2, int size2);
+int		str_append_str(char **s1, char *s2);
+void	*ft_memmove(void *dest, const void *src, int n);
 
 #endif
